loki.batch: Move ItemFactory into separate module
