+++ loki/batch/__init__.py
@@ -16,6 +16,7 @@ configuration utilities for large call tree traversals.
 
 from loki.batch.configure import * # noqa
 from loki.batch.item import * # noqa
+from loki.batch.item_factory import * # noqa
 from loki.batch.pipeline import * # noqa
 from loki.batch.scheduler import * # noqa
 from loki.batch.sfilter import * # noqa
